fix: certificate validation
certificate validation overrided from core to avoid medical rules failures
